reorg c14n/catalog/stream coverage tests into themed subtests
